Adapt protocolmock to changes in cryptoprotoocls and rename the module to protocoldefinition.
