+++ protocoldefinition/src/main/java/de/upb/crypto/incentive/protocoldefinition/creditearn/CreditResponse.java
@@ -1,14 +1,15 @@
-package de.upb.crypto.incentive.protocolmock.credit;
+package de.upb.crypto.incentive.protocoldefinition.creditearn;
 
-import de.upb.crypto.incentive.protocolmock.model.Token;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.util.UUID;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class CreditResponse {
-    private int id;
-    private Token token;
+    private UUID id;
+    private String serializedCreditResponse;
 }
