+++ protocoldefinition/src/main/java/de/upb/crypto/incentive/protocoldefinition/creditearn/EarnRequest.java
@@ -1,0 +1,17 @@
+package de.upb.crypto.incentive.protocoldefinition.creditearn;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.UUID;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class EarnRequest {
+    private UUID id;
+    private String serializedEarnRequest;
+    private long earnAmount;
+    private UUID basketId;
+}
