+++ protocoldefinition/src/main/java/de/upb/crypto/incentive/protocoldefinition/deduct/DeductRequest.java
@@ -1,6 +1,6 @@
-package de.upb.crypto.incentive.protocolmock.deduct;
+package de.upb.crypto.incentive.protocoldefinition.deduct;
 
-import de.upb.crypto.incentive.protocolmock.model.Token;
+import de.upb.crypto.incentive.protocoldefinition.model.Token;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
